separate custom cluster into __init__
