+++ zhaquirks/xiaomi/aqara/aqara_z1_pro_single.py
@@ -3,7 +3,6 @@
 import logging
 import sys
 from zigpy.profiles import zha
-from zigpy.quirks import CustomDevice, CustomCluster
 from zigpy.zcl.clusters.general import (
     AnalogInput,
     Basic,
@@ -23,7 +22,6 @@ from zhaquirks.const import (
     MODELS_INFO,
     OUTPUT_CLUSTERS,
     PROFILE_ID,
-    ZHA_SEND_EVENT,
 )
 from zhaquirks.xiaomi import (
     LUMI,
@@ -33,6 +31,7 @@ from zhaquirks.xiaomi import (
     MeteringCluster,
     OnOffCluster,
     XiaomiCustomDevice,
+    AqaraZ1ProManufacturerSpecificCluster
 )
 from zhaquirks.xiaomi.aqara.opple_remote import MultistateInputCluster
 
@@ -49,84 +48,6 @@ _LOGGER.addHandler(console_handler)
 
 # Log at module level to verify the file is being loaded
 _LOGGER.debug("AqaraZ1ProSingleRockerSwitch quirk module is being loaded! ZHA Profile ID: 0x%04x, Device Type: 0x%04x", zha.PROFILE_ID, zha.DeviceType.ON_OFF_SWITCH)
-
-class AqaraZ1ProManufacturerSpecificCluster(CustomCluster):
-    """Custom cluster for Aqara Z1 Pro manufacturer specific events."""
-    
-    cluster_id = 0xfcc0
-    
-    # Attribute IDs
-    ATTR_SLIDER_ACTION = 0x028C  # 652
-    ATTR_SLIDE_TIME = 0x0231     # 561
-    ATTR_SLIDE_SPEED = 0x0232    # 562
-    ATTR_SLIDE_RELATIVE_DISPLACEMENT = 0x0233  # 563
-    ATTR_SLIDE_TIME_DELTA = 0x0301  # 769
-
-    # ATTR_DEVICE_ID_SHADE = 0x0200 # 512
-    # ATTR_LOCK_RELAY = 0x0285 # 645
-    # ATTR_SWITCH_MODE = 0x0004 # 4
-    # ATTR_POWER_ON_BEHAVIOR = 0x0517 # 1303
-    # ATTR_CLICK_MODE = 0x0125 # 293
-    
-    # Action mapping
-    ACTION_MAPPING = {
-        1: "slider_single",
-        2: "slider_double",
-        3: "slider_hold",
-        4: "slider_up",
-        5: "slider_down",
-    }
-    
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._attr_id = self.ATTR_SLIDER_ACTION
-        _LOGGER.debug("AqaraZ1ProManufacturerSpecificCluster initialized for device %s", self._endpoint.device.ieee)
-        
-    def _update_attribute(self, attrid, value):
-        """Handle attribute updates."""
-        
-        # Store all attributes for the event
-        if not hasattr(self, "_manufacturer_attrs"):
-            self._manufacturer_attrs = {}
-        
-        # Update the attribute value
-        self._manufacturer_attrs[attrid] = value
-        
-        # If this is the slider action attribute, send the event
-        if attrid == self.ATTR_SLIDER_ACTION:
-            # Get the action name from the mapping
-            action = self.ACTION_MAPPING.get(value, f"slider_unknown_{value}")
-            _LOGGER.debug("AqaraZ1ProManufacturerSpecificCluster detected action: %s (value: %s)", action, value)
-            
-            # Prepare the event data
-            event_data = {
-                "action": action,
-                "value": value,
-                "slide_time": self._manufacturer_attrs.get(self.ATTR_SLIDE_TIME),
-                "slide_speed": self._manufacturer_attrs.get(self.ATTR_SLIDE_SPEED),
-                "slide_relative_displacement": self._manufacturer_attrs.get(self.ATTR_SLIDE_RELATIVE_DISPLACEMENT),
-                "slide_time_delta": self._manufacturer_attrs.get(self.ATTR_SLIDE_TIME_DELTA),
-            }
-            
-            _LOGGER.debug("AqaraZ1ProManufacturerSpecificCluster sending event data: %s", event_data)
-            
-            # Send the event
-            self.listener_event(
-                ZHA_SEND_EVENT,
-                action,
-                event_data,
-            )
-            
-            # Also send a generic slider event for easier automation
-            self.listener_event(
-                ZHA_SEND_EVENT,
-                "slider_event",
-                event_data,
-            )
-            _LOGGER.debug("AqaraZ1ProManufacturerSpecificCluster events sent successfully")
-
-        _LOGGER.debug("AqaraZ1ProManufacturerSpecificCluster attribute update: attrid=0x%04x, value=%s", attrid, value)
-        super()._update_attribute(attrid, value)
 
 class AqaraZ1ProSingleRockerSwitch(XiaomiCustomDevice):
     """Aqara Z1 Pro Single Rocker Switch"""
